requirements.txt
openai==1.16.2
python-dotenv==1.0.1
requests
selenium
opencv-python
ipython
